blick-mini-app/requirements.txt
flask==3.0.3
httpx==0.27.0
python-dotenv==1.0.1
gunicorn==22.0.0
